# specs/019-vendor-credit-alert/contracts/vendor-credit-alert.openapi.yaml
openapi: 3.1.0
info:
  title: Vendor Credit Alert API Impact
  version: 0.1.0
  description: >-
    Agent-facing contract updates for decoupled vendor credit alerting. Successful
    invoice create and approve responses keep their existing business fields, add a
    top-level `credit_check_id`, and schedule a post-response background evaluation. Agents can query the background
    check by identifier and can retrieve the active alert through vendor exposure.
servers:
  - url: http://localhost:8000
paths:
  /invoices:
    post:
      summary: Register an invoice and schedule a post-response credit exposure check
      operationId: createInvoiceWithCreditCheck
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceRequest"
      responses:
        "201":
          description: >-
            Invoice registered successfully. Existing business fields are unchanged.
            The response includes one deterministic `credit_check_id`. A background credit exposure check is scheduled only
            for the first logical success of the idempotent request.
          headers:
            X-Correlation-ID:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceCreateResponse"
              examples:
                created:
                  value:
                    invoice_id: INV-ALERT-100
                    vendor_id: V-100
                    purchase_order_id: PO-ALERT-100
                    invoice_number: INV-ALERT-100
                    invoice_amount: "3500.00"
                    status: PENDING
                    last_match_outcome: NONE
                    next_action: REQUEST_MATCH
                    credit_check_id: CHK-123
        "404":
          description: Vendor or purchase order was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Duplicate vendor invoice reference, vendor and purchase-order mismatch, or conflicting idempotency-key reuse
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Invoice-registration input is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /invoices/{invoiceId}/approve:
    post:
      summary: Approve a matched invoice and schedule a post-response credit exposure check
      operationId: approveInvoiceWithCreditCheck
      parameters:
        - $ref: "#/components/parameters/InvoiceId"
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: >-
            Invoice approved successfully. Existing business fields are unchanged.
            The response includes one deterministic `credit_check_id`. A background credit exposure check is scheduled only
            for the first logical success of the idempotent request.
          headers:
            X-Correlation-ID:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceApprovalResponse"
              examples:
                approved:
                  value:
                    invoice_id: INV-ALERT-100
                    invoice_status: APPROVED
                    purchase_order_id: PO-ALERT-100
                    generated_gl_entries:
                      - gl_entry_id: GL-1
                        account_code: AP_CONTROL
                        debit: "3500.00"
                        credit: "0.00"
                      - gl_entry_id: GL-2
                        account_code: EXPENSE_BUILDING_SUPPLY
                        debit: "0.00"
                        credit: "3500.00"
                    approved_at: "2026-05-25T09:03:00Z"
                    next_action: MARK_PAID
                    credit_check_id: CHK-456
        "404":
          description: Invoice not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Invoice is in the wrong state for approval or the idempotency key conflicts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Retryable dependency failure prevented approval completion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /credit-checks/{creditCheckId}:
    get:
      summary: Query the outcome of a specific background credit check
      operationId: getCreditCheckStatus
      parameters:
        - $ref: "#/components/parameters/CreditCheckId"
        - $ref: "#/components/parameters/CorrelationId"
      responses:
        "200":
          description: Credit check state returned successfully
          headers:
            X-Correlation-ID:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditCheckStatusResponse"
              examples:
                pending:
                  value:
                    status: PENDING
                    breached: null
                    alert_id: null
                completedWithoutBreach:
                  value:
                    status: COMPLETED
                    breached: false
                    alert_id: null
                completedWithBreach:
                  value:
                    status: COMPLETED
                    breached: true
                    alert_id: ALERT-456
        "404":
          description: Credit check was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Temporary infrastructure failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /vendors/{vendorId}/exposure:
    get:
      summary: Retrieve current AP exposure for a vendor, including any active credit alert
      operationId: getVendorExposureWithCreditAlert
      parameters:
        - $ref: "#/components/parameters/VendorId"
        - $ref: "#/components/parameters/CorrelationId"
      responses:
        "200":
          description: Exposure summary returned successfully
          headers:
            X-Correlation-ID:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorExposureResponse"
              examples:
                withoutAlert:
                  value:
                    vendor_id: V-100
                    vendor_name: ACME Building Supply
                    as_of_timestamp: "2026-05-25T09:00:00Z"
                    outstanding_total_amount: "1750.00"
                    open_invoice_count: 2
                    included_invoice_statuses:
                      - PENDING
                      - MATCHED
                      - APPROVED
                withAlert:
                  value:
                    vendor_id: V-100
                    vendor_name: ACME Building Supply
                    as_of_timestamp: "2026-05-25T09:05:00Z"
                    outstanding_total_amount: "2300.00"
                    open_invoice_count: 3
                    included_invoice_statuses:
                      - PENDING
                      - MATCHED
                      - APPROVED
                    active_credit_alert:
                      alert_id: ALERT-456
                      vendor_id: V-100
                      credit_check_id: CHK-456
                      triggering_invoice_id: INV-ALERT-100
                      outstanding_amount: "2300.00"
                      credit_limit: "2000.00"
                      percentage_consumed: "115.00"
                      breached_at: "2026-05-25T09:04:59Z"
                      advisory_only: true
        "404":
          description: Vendor was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Temporary infrastructure failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  parameters:
    InvoiceId:
      name: invoiceId
      in: path
      required: true
      schema:
        type: string
    CreditCheckId:
      name: creditCheckId
      in: path
      required: true
      schema:
        type: string
    VendorId:
      name: vendorId
      in: path
      required: true
      schema:
        type: string
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: true
      schema:
        type: string
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
  headers:
    CorrelationId:
      description: Correlation identifier echoed from the request
      schema:
        type: string
  schemas:
    CreateInvoiceRequest:
      type: object
      required:
        - vendor_id
        - purchase_order_id
        - invoice_number
        - invoice_amount
      properties:
        vendor_id:
          type: string
        purchase_order_id:
          type: string
        invoice_number:
          type: string
        invoice_amount:
          type: string
    InvoiceCreateResponse:
      type: object
      required:
        - invoice_id
        - vendor_id
        - purchase_order_id
        - invoice_number
        - invoice_amount
        - status
        - last_match_outcome
        - next_action
        - credit_check_id
      properties:
        invoice_id:
          type: string
        vendor_id:
          type: string
        purchase_order_id:
          type: string
        invoice_number:
          type: string
        invoice_amount:
          type: string
        status:
          type: string
        last_match_outcome:
          type: string
        next_action:
          type: string
        credit_check_id:
          type: string
    InvoiceApprovalResponse:
      type: object
      required:
        - invoice_id
        - invoice_status
        - purchase_order_id
        - generated_gl_entries
        - approved_at
        - next_action
        - credit_check_id
      properties:
        invoice_id:
          type: string
        invoice_status:
          type: string
          enum:
            - APPROVED
        purchase_order_id:
          type: string
        generated_gl_entries:
          type: array
          minItems: 2
          maxItems: 2
          items:
            $ref: "#/components/schemas/GLEntrySummary"
        approved_at:
          type: string
          format: date-time
        next_action:
          type: string
          enum:
            - MARK_PAID
        credit_check_id:
          type: string
    CreditCheckStatusResponse:
      type: object
      required:
        - status
        - breached
        - alert_id
      properties:
        status:
          type: string
          enum:
            - PENDING
            - COMPLETED
        breached:
          type:
            - boolean
            - "null"
        alert_id:
          type:
            - string
            - "null"
    GLEntrySummary:
      type: object
      required:
        - gl_entry_id
        - account_code
        - debit
        - credit
      properties:
        gl_entry_id:
          type: string
        account_code:
          type: string
        debit:
          type: string
        credit:
          type: string
    CreditAlertSummary:
      type: object
      required:
        - alert_id
        - vendor_id
        - credit_check_id
        - triggering_invoice_id
        - outstanding_amount
        - credit_limit
        - percentage_consumed
        - breached_at
        - advisory_only
      properties:
        alert_id:
          type: string
        vendor_id:
          type: string
        credit_check_id:
          type: string
        triggering_invoice_id:
          type: string
        outstanding_amount:
          type: string
        credit_limit:
          type: string
        percentage_consumed:
          type: string
        breached_at:
          type: string
          format: date-time
        advisory_only:
          type: boolean
          enum:
            - true
    VendorExposureResponse:
      type: object
      required:
        - vendor_id
        - vendor_name
        - as_of_timestamp
        - outstanding_total_amount
        - open_invoice_count
        - included_invoice_statuses
      properties:
        vendor_id:
          type: string
        vendor_name:
          type: string
        as_of_timestamp:
          type: string
          format: date-time
        outstanding_total_amount:
          type: string
        open_invoice_count:
          type: integer
          minimum: 0
        included_invoice_statuses:
          type: array
          items:
            type: string
        active_credit_alert:
          oneOf:
            - $ref: "#/components/schemas/CreditAlertSummary"
            - type: "null"
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - category
            - retryable
            - message
            - correlation_id
          properties:
            code:
              type: string
              enum:
                - VENDOR_NOT_FOUND
                - PURCHASE_ORDER_NOT_FOUND
                - INVOICE_NOT_FOUND
                - INVOICE_INVALID_STATE
                - IDEMPOTENCY_KEY_CONFLICT
                - DUPLICATE_VENDOR_INVOICE_REFERENCE
                - CREDIT_CHECK_NOT_FOUND
                - DEPENDENCY_TEMPORARILY_UNAVAILABLE
            category:
              type: string
              enum:
                - business
                - infrastructure
            retryable:
              type: boolean
            message:
              type: string
            correlation_id:
              type: string